Create new project action in searcher. (#1566)
