+++ src/rust/ide/src/tests.rs
@@ -26,11 +26,10 @@ fn failure_to_open_project_is_reported() {
     let transport   = MockTransport::new();
     let mut fixture = TestWithMockedTransport::set_up(&transport);
     fixture.run_test(async move {
-        let logger          = Logger::new("test");
         let project_manager = Rc::new(project_manager::Client::new(transport));
         executor::global::spawn(project_manager.runner());
         let name        = ProjectName(crate::constants::DEFAULT_PROJECT_NAME.to_owned());
-        let initializer = ide::initializer::WithProjectManager::new(logger,project_manager,name);
+        let initializer = ide::initializer::WithProjectManager::new(project_manager,name);
         let result      = initializer.initialize_project_model().await;
         result.expect_err("Error should have been reported.");
     });
